+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add public API for crash-on-launch detection
   [#1157](https://github.com/bugsnag/bugsnag-android/pull/1157)
   [#1159](https://github.com/bugsnag/bugsnag-android/pull/1159)
+  [#1165](https://github.com/bugsnag/bugsnag-android/pull/1165)
 
 ## 5.7.0 (2021-02-18)